fix: make shellcheck happy
Cleaned up a bunch of bad shell script which `shellcheck` rightfully
complained about. Was also forced to disable a few checks, most
selectively and 2 globally, which are mandatory for shlib to work. Most
of these would be valid complaints in any environment in which a shell
script was used as a stand-alone application, but they make
little-to-no-sense from a shell library perspective.