feat: initial support for subcommands
Adding support for `shlib <command>` along side `shlib <file>`.

The core idea here is that we have `shlib <arg>`, and should `<arg>`
start with one of `/`, `./`, or `../`, then we consider it a file,
else we consider it a sub-command.

This will allow us to support sub-commands as part of the `shlib`
framework. E.g. `shlib test`, `shlib install`, `shlib doc`, etc.